Move DataErrorNotifier up the root namespace
There no longer is need of a ToFedora module and this class was used outside of the module anyway
