+++ app/services/cocina/to_fedora/content_metadata_generator.rb
@@ -178,7 +178,7 @@ module Cocina
       end
 
       def notifier
-        @notifier ||= FromFedora::DataErrorNotifier.new(druid: druid)
+        @notifier ||= DataErrorNotifier.new(druid: druid)
       end
 
       def type_for(cocina_fileset)
